# frontend/src/types/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "人生草稿本 API",
    "description": "Life Draft Book — Interactive narrative game API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/games/{game_id}/choice": {
      "post": {
        "tags": [
          "Gameplay"
        ],
        "summary": "Make Choice",
        "description": "Process a player choice via SSE streaming (story continuation + post-processing).\n\nSupports reconnection via Last-Event-ID header for mobile network resilience.",
        "operationId": "make_choice_api_games__game_id__choice_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MakeChoiceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/games/{game_id}/custom-choice": {
      "post": {
        "tags": [
          "Gameplay"
        ],
        "summary": "Make Custom Choice",
        "description": "Process a custom player choice via SSE streaming.\n\nSupports reconnection via Last-Event-ID header for mobile network resilience.",
        "operationId": "make_custom_choice_api_games__game_id__custom_choice_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomChoiceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/games/{game_id}/choice-sync": {
      "post": {
        "tags": [
          "Gameplay"
        ],
        "summary": "Make Choice Sync",
        "description": "Process a player choice (non-streaming fallback for mobile).",
        "operationId": "make_choice_sync_api_games__game_id__choice_sync_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MakeChoiceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/games/{game_id}/custom-choice-sync": {
      "post": {
        "tags": [
          "Gameplay"
        ],
        "summary": "Make Custom Choice Sync",
        "description": "Process a custom player choice (non-streaming fallback for mobile).",
        "operationId": "make_custom_choice_sync_api_games__game_id__custom_choice_sync_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomChoiceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/games/{game_id}/summary": {
      "post": {
        "tags": [
          "Gameplay"
        ],
        "summary": "Generate Summary",
        "description": "Generate life summary based on complete story history.",
        "operationId": "generate_summary_api_games__game_id__summary_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSummaryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomChoiceRequest": {
        "properties": {
          "custom_text": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Custom Text"
          }
        },
        "type": "object",
        "required": [
          "custom_text"
        ],
        "title": "CustomChoiceRequest"
      },
...
      "GenerateSummaryRequest": {
        "properties": {
          "weeks": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Weeks",
            "default": 10
          }
        },
        "type": "object",
        "title": "GenerateSummaryRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MakeChoiceRequest": {
        "properties": {
          "option_index": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Option Index"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          }
        },
        "type": "object",
        "required": [
          "option_index"
        ],
        "title": "MakeChoiceRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}